+++ scripts/pypsa-de/additional_functionality.py
@@ -384,6 +384,9 @@ def add_national_co2_budgets(n, snakemake, national_co2_budgets, investment_year
             logger.error(
                 f"CO2 budget for countries other than `DE` is not yet supported. Found country {ct}. Please check the config file."
             )
+            raise NotImplementedError(
+                f"CO2 budget for countries other than `DE` is not yet supported. Found country {ct}. Please check the config file."
+            )
 
         limit = co2_total_totals[ct] * national_co2_budgets[ct][investment_year]
         logger.info(