+++ scripts/pypsa-de/build_existing_chp_de.py
@@ -12,7 +12,12 @@ import logging
 import geopandas as gpd
 import pandas as pd
 
-from scripts._helpers import configure_logging, mock_snakemake
+from scripts._helpers import (
+    configure_logging,
+    mock_snakemake,
+    set_scenario_config,
+    update_config_from_wildcards,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -265,6 +270,9 @@ if __name__ == "__main__":
         )
 
     configure_logging(snakemake)
+    set_scenario_config(snakemake)
+    update_config_from_wildcards(snakemake.config, snakemake.wildcards)
+
     logger.info("Retrieving and cleaning CHP data from BNetzA")
     biomass = pd.read_csv(
         snakemake.input.mastr_biomass, dtype={"Postleitzahl": str}, low_memory=False