+++ scripts/pypsa-de/modify_cost_data.py
@@ -4,7 +4,12 @@ import re
 
 import pandas as pd
 
-from scripts._helpers import configure_logging, mock_snakemake
+from scripts._helpers import (
+    configure_logging,
+    mock_snakemake,
+    set_scenario_config,
+    update_config_from_wildcards,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -48,6 +53,8 @@ if __name__ == "__main__":
             run="KN2045_Mix",
         )
     configure_logging(snakemake)
+    set_scenario_config(snakemake)
+    update_config_from_wildcards(snakemake.config, snakemake.wildcards)
 
     # read in cost data from technology-data library
     costs = os.path.join(