+++ scripts/pypsa-de/modify_district_heat_share.py
@@ -25,7 +25,12 @@ import logging
 import geopandas as gpd
 import pandas as pd
 
-from scripts._helpers import mock_snakemake
+from scripts._helpers import (
+    configure_logging,
+    mock_snakemake,
+    set_scenario_config,
+    update_config_from_wildcards,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -107,7 +112,9 @@ if __name__ == "__main__":
             run="KN2045_Mix",
         )
 
-    logging.basicConfig(level=snakemake.config["logging"]["level"])
+    configure_logging(snakemake)
+    set_scenario_config(snakemake)
+    update_config_from_wildcards(snakemake.config, snakemake.wildcards)
     logger.info("Updating district heating shares with egon data")
 
     heat_techs = gpd.read_file(snakemake.input.heating_technologies_nuts3)