+++ scripts/pypsa-de/plot_hydrogen_network_incl_kernnetz.py
@@ -9,12 +9,24 @@ selected other infrastructure.
 import logging
 
 import geopandas as gpd
+import matplotlib
+
+matplotlib.use("Agg")
 import matplotlib.pyplot as plt
 import pandas as pd
 import pypsa
-from pypsa.plot import add_legend_circles, add_legend_lines, add_legend_patches
+from pypsa.plot.maps.static import (
+    add_legend_circles,
+    add_legend_lines,
+    add_legend_patches,
+)
 
-from scripts._helpers import configure_logging, mock_snakemake, set_scenario_config
+from scripts._helpers import (
+    configure_logging,
+    mock_snakemake,
+    set_scenario_config,
+    update_config_from_wildcards,
+)
 from scripts.make_summary import assign_locations
 from scripts.plot_power_network import load_projection
 
@@ -211,7 +223,7 @@ def plot_h2_map(n, regions):
 
     bus_colors = {"H2 Electrolysis": "#ff29d9", "H2 Fuel Cell": "#805394"}
 
-    n.plot(
+    n.plot.map(
         geomap=True,
         bus_sizes=bus_sizes,
         bus_colors=bus_colors,
@@ -222,7 +234,7 @@ def plot_h2_map(n, regions):
         **map_opts,
     )
 
-    n.plot(
+    n.plot.map(
         geomap=True,
         bus_sizes=0,
         link_colors=color_retrofit,
@@ -232,7 +244,7 @@ def plot_h2_map(n, regions):
         **map_opts,
     )
 
-    n.plot(
+    n.plot.map(
         geomap=True,
         bus_sizes=0,
         link_colors=color_kern,
@@ -339,6 +351,7 @@ if __name__ == "__main__":
 
     configure_logging(snakemake)
     set_scenario_config(snakemake)
+    update_config_from_wildcards(snakemake.config, snakemake.wildcards)
 
     n = pypsa.Network(snakemake.input.network)