+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+- Simplified IIASA database download
 - Bugfix: Enforce stricter H2 derivative import limit to avoid that exports of one type of derivative compensate for imports of another
 - Added an option to source mobility demand from UBA MWMS (Projektionsbericht 2025) for the years 2025-2035
 - Renamed functions and script for exogenous mobility demand